backend/requirements.txt
# Django
Django>=4.2,<5.0

# REST API
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# CORS headers for React frontend
django-cors-headers>=4.3.0

# Image handling
Pillow>=10.0.0

# Environment variables (for production)
python-dotenv>=1.0.0

# Password validation
django-password-validators>=1.7.0